src/scripts/archive/comic-all-in-one/requirements.txt
Pillow>=9.0.0
send2trash>=1.8.0
questionary>=1.10.0
pyperclip>=1.8.2
pillow-avif>=1.3.1
pillow-jxl>=0.3.0
numpy>=1.21.0 